// frontend/src/backend.d.ts
export enum ChartTimeframe {
    fourHours = "fourHours",
    oneHour = "oneHour",
    sevenDays = "sevenDays",
    oneYear = "oneYear",
    eightHours = "eightHours",
    fifteenMinutes = "fifteenMinutes",
    oneMonth = "oneMonth",
    oneDay = "oneDay"
}